refactoring to get ready for filter embeddings
