+++ modisco/seqlet_embedding/__init__.py
@@ -1,0 +1,1 @@
+from . import gapped_kmer
